Changed format of note for missing data to include more details than just value.
